// generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  ActPolicyAndSchemeContentDynamicZoneInput: { input: any; output: any; }
  AllMediaContentDynamicZoneInput: { input: any; output: any; }
  AllPackageContentDynamicZoneInput: { input: any; output: any; }
  BestTimeContentDynamicZoneInput: { input: any; output: any; }
  CityDetailContentDynamicZoneInput: { input: any; output: any; }
  ContactUsContentDynamicZoneInput: { input: any; output: any; }
  /** A date string, such as 2007-12-03, compliant with the `full-date` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: { input: any; output: any; }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: { input: any; output: any; }
  DestinationContentDynamicZoneInput: { input: any; output: any; }
  EmployeeCornerContentDynamicZoneInput: { input: any; output: any; }
  ExploreAllContentDynamicZoneInput: { input: any; output: any; }
  ExploreContentDynamicZoneInput: { input: any; output: any; }
  FilmShootingContentDynamicZoneInput: { input: any; output: any; }
  FilmistanContentDynamicZoneInput: { input: any; output: any; }
  ForeignTouristContentDynamicZoneInput: { input: any; output: any; }
  HeritagePropertyContentDynamicZoneInput: { input: any; output: any; }
  HomeHomeDynamicZoneInput: { input: any; output: any; }
  HowToGetThereContentDynamicZoneInput: { input: any; output: any; }
  IhmsAndFciContentDynamicZoneInput: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
  /** The `BigInt` scalar type represents non-fractional signed whole numeric values. */
  Long: { input: any; output: any; }
  LuxuryOnWheelContentDynamicZoneInput: { input: any; output: any; }
  PlaceDetailContentDynamicZoneInput: { input: any; output: any; }
  RajasthanTravelGuideContentDynamicZoneInput: { input: any; output: any; }
  RegistrationForTravelAgencyContentDynamicZoneInput: { input: any; output: any; }
  SeeAllPlaceContentDynamicZoneInput: { input: any; output: any; }
  SuggestedItinerieContentDynamicZoneInput: { input: any; output: any; }
  TravelDairyContentDynamicZoneInput: { input: any; output: any; }
  TravelDeskBannerDynamicZoneInput: { input: any; output: any; }
  /** The `Upload` scalar type represents a file upload. */
  Upload: { input: any; output: any; }
};
// ...
export type ComponentPlaceDetail3DVideo = {
  __typename?: 'ComponentPlaceDetail3DVideo';
  id: Scalars['ID']['output'];
  threeDvideo?: Maybe<UploadFileRelationResponseCollection>;
};
// ...
export type ComponentPlaceDetailContent = {
  __typename?: 'ComponentPlaceDetailContent';
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  image?: Maybe<UploadFileEntityResponse>;
  title?: Maybe<Scalars['String']['output']>;
};

export type ComponentPlaceDetailContentFiltersInput = {
  and?: InputMaybe<Array<InputMaybe<ComponentPlaceDetailContentFiltersInput>>>;
  description?: InputMaybe<StringFilterInput>;
  not?: InputMaybe<ComponentPlaceDetailContentFiltersInput>;
  or?: InputMaybe<Array<InputMaybe<ComponentPlaceDetailContentFiltersInput>>>;
  title?: InputMaybe<StringFilterInput>;
};

export type ComponentPlaceDetailDescription = {
  __typename?: 'ComponentPlaceDetailDescription';
  id: Scalars['ID']['output'];
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type ComponentPlaceDetailDynamicprice = {
  __typename?: 'ComponentPlaceDetailDynamicprice';
  cats?: Maybe<Array<Maybe<ComponentPlaceDetailDpCat>>>;
  id: Scalars['ID']['output'];
};
// ...
export type ComponentPlaceDetailExhibit = {
  __typename?: 'ComponentPlaceDetailExhibit';
  content?: Maybe<Array<Maybe<ComponentPlaceDetailContent>>>;
  id: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
  slug?: Maybe<Scalars['String']['output']>;
};
// ...
export type ComponentPlaceDetailHeader = {
  __typename?: 'ComponentPlaceDetailHeader';
  id: Scalars['ID']['output'];
};

export type ComponentPlaceDetailImages = {
  __typename?: 'ComponentPlaceDetailImages';
  id: Scalars['ID']['output'];
};
// ...
export type ComponentPlaceDetailPlaceAttraction = {
  __typename?: 'ComponentPlaceDetailPlaceAttraction';
  card?: Maybe<Array<Maybe<ComponentPlaceDetailCard>>>;
  id: Scalars['ID']['output'];
  title1?: Maybe<Scalars['String']['output']>;
};
// ...
export type ComponentPlaceDetailPlaceGallery = {
  __typename?: 'ComponentPlaceDetailPlaceGallery';
  id: Scalars['ID']['output'];
  image?: Maybe<UploadFileRelationResponseCollection>;
};
// ...
export type ComponentPlaceDetailPlaceTime = {
  __typename?: 'ComponentPlaceDetailPlaceTime';
  card?: Maybe<Array<Maybe<ComponentPlaceDetailTickets>>>;
  id: Scalars['ID']['output'];
  tabName: Scalars['String']['output'];
};


export type ComponentPlaceDetailPlaceTimeCardArgs = {
  filters?: InputMaybe<ComponentPlaceDetailTicketsFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type ComponentPlaceDetailPlaceexhibit = {
  __typename?: 'ComponentPlaceDetailPlaceexhibit';
  exhibit?: Maybe<Array<Maybe<ComponentPlaceDetailExhibit>>>;
  id: Scalars['ID']['output'];
};
// ...
export type ComponentPlaceDetailPlaceothers = {
  __typename?: 'ComponentPlaceDetailPlaceothers';
  card?: Maybe<Array<Maybe<ComponentPlaceDetailTickets>>>;
  id: Scalars['ID']['output'];
  title1?: Maybe<Scalars['String']['output']>;
  title2?: Maybe<Scalars['String']['output']>;
};


export type ComponentPlaceDetailPlaceothersCardArgs = {
  filters?: InputMaybe<ComponentPlaceDetailTicketsFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type ComponentPlaceDetailPlaceoverview = {
  __typename?: 'ComponentPlaceDetailPlaceoverview';
  id: Scalars['ID']['output'];
  overview?: Maybe<ComponentPlaceDetailOverview>;
};

export type ComponentPlaceDetailPlacestovisit = {
  __typename?: 'ComponentPlaceDetailPlacestovisit';
  description?: Maybe<Scalars['String']['output']>;
  icon?: Maybe<UploadFileEntityResponse>;
  id: Scalars['ID']['output'];
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type PlaceDetail = {
  __typename?: 'PlaceDetail';
  categories?: Maybe<Array<Maybe<ComponentPlaceDetailCategory>>>;
  content?: Maybe<Array<Maybe<PlaceDetailContentDynamicZone>>>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  place?: Maybe<PlaceEntityResponse>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  slug: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};
// ...
export type PlaceDetailEntity = {
  __typename?: 'PlaceDetailEntity';
  attributes?: Maybe<PlaceDetail>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type PlaceDetailEntityResponse = {
  __typename?: 'PlaceDetailEntityResponse';
  data?: Maybe<PlaceDetailEntity>;
};

export type PlaceDetailEntityResponseCollection = {
  __typename?: 'PlaceDetailEntityResponseCollection';
  data: Array<PlaceDetailEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type UploadFileEntity = {
  __typename?: 'UploadFileEntity';
  attributes?: Maybe<UploadFile>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type UploadFileEntityResponse = {
  __typename?: 'UploadFileEntityResponse';
  data?: Maybe<UploadFileEntity>;
};
// ...
export type FetchExhibitDetailQueryVariables = Exact<{
  slug?: InputMaybe<Scalars['String']['input']>;
  exhibitSlug?: InputMaybe<Scalars['String']['input']>;
}>;


export type FetchExhibitDetailQuery = { __typename?: 'Query', placeDetails?: { __typename?: 'PlaceDetailEntityResponseCollection', data: Array<{ __typename?: 'PlaceDetailEntity', attributes?: { __typename?: 'PlaceDetail', content?: Array<
          | { __typename?: 'ComponentPlaceDetail3DVideo' }
          | { __typename?: 'ComponentPlaceDetailDescription' }
          | { __typename?: 'ComponentPlaceDetailDynamicprice' }
          | { __typename?: 'ComponentPlaceDetailHeader' }
          | { __typename?: 'ComponentPlaceDetailImages' }
          | { __typename?: 'ComponentPlaceDetailPlaceAttraction' }
          | { __typename?: 'ComponentPlaceDetailPlaceGallery' }
          | { __typename?: 'ComponentPlaceDetailPlaceTime' }
          | { __typename: 'ComponentPlaceDetailPlaceexhibit', id: string, exhibit?: Array<{ __typename?: 'ComponentPlaceDetailExhibit', name?: string | null, slug?: string | null, content?: Array<{ __typename?: 'ComponentPlaceDetailContent', title?: string | null, description?: string | null, image?: { __typename?: 'UploadFileEntityResponse', data?: { __typename?: 'UploadFileEntity', id?: string | null, attributes?: { __typename?: 'UploadFile', url: string, alternativeText?: string | null } | null } | null } | null } | null> | null } | null> | null }
          | { __typename?: 'ComponentPlaceDetailPlaceothers' }
          | { __typename?: 'ComponentPlaceDetailPlaceoverview' }
          | { __typename?: 'ComponentPlaceDetailPlacestovisit' }
          | { __typename?: 'ComponentPlaceDetailPlacetickets' }
          | { __typename?: 'Error' }
         | null> | null } | null }> } | null };
// ...
export const FetchExhibitDetailDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"FetchExhibitDetail"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"slug"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"exhibitSlug"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"placeDetails"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"filters"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"slug"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"eq"},"value":{"kind":"Variable","name":{"kind":"Name","value":"slug"}}}]}}]}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"data"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"attributes"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"content"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"InlineFragment","typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"ComponentPlaceDetailPlaceexhibit"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}},{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"exhibit"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"pagination"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"limit"},"value":{"kind":"IntValue","value":"999"}}]}},{"kind":"Argument","name":{"kind":"Name","value":"filters"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"slug"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"eq"},"value":{"kind":"Variable","name":{"kind":"Name","value":"exhibitSlug"}}}]}}]}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"slug"}},{"kind":"Field","name":{"kind":"Name","value":"content"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"image"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"data"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"attributes"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"url"}},{"kind":"Field","name":{"kind":"Name","value":"alternativeText"}}]}}]}}]}}]}}]}}]}}]}}]}}]}}]}}]}}]} as unknown as DocumentNode<FetchExhibitDetailQuery, FetchExhibitDetailQueryVariables>;